% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{mergeStages}
\alias{mergeStages}
\title{Misc functions}
\usage{
mergeStages(patient.stage_event.pathologic_stage)
}
\description{
Misc functions
}
\seealso{
Other RTCGA.tools: \code{\link{calculate_cooccurence}};
  \code{\link{clinicalStageSurvival}};
  \code{\link{mergeTCGA}}
}

