% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{clinicalStageSurvival}
\alias{clinicalStageSurvival}
\title{Plot functions}
\usage{
clinicalStageSurvival(clinical, xlims = c(0, 2000),
  title = "Marks show times with censoring")
}
\description{
Plot functions
}
\examples{
\dontrun{
library(RTCGA.clinical)
clinicalStageSurvival(LUAD.clinical, xlims = c(0,2000))

library(RTCGA.rnaseq)
rnaseqBiplot(cohorts = c("ACC", "CHOL", "GBM", "PCPG", "UVM"))

library(RTCGA.mutations)
mutationsBox(c("BRCA", "HNSC", "LUSC", "PRAD", "STES"),
 "TP53", "MDM2")
 mutationsBox(c("BRCA", "HNSC", "LUSC", "PRAD", "STES"),
  "TP53", "ETF1")
}
}
\seealso{
Other RTCGA.tools: \code{\link{calculate_cooccurence}};
  \code{\link{mergeStages}}; \code{\link{mergeTCGA}}
}

